requirements-images.txt
# Requirements for image extraction script
requests>=2.28.0
beautifulsoup4>=4.11.0
Pillow>=9.0.0
lxml>=4.9.0
